# api/openapi.yaml
components:
  schemas:
    Article:
      properties:
        author:
          type: string
        body:
          items:
            type: string
          type: array
        externalUrl:
          type: string
        fetchedAt:
          format: date-time
          type: string
        headline:
          type: string
        id:
          type: string
        publishedAt:
          format: date-time
          type: string
        readingMinutes:
          format: int32
          type: integer
        snippet:
          type: string
        sourceId:
          type: string
        topicId:
          type: string
      type: object
...
    LoginRequest:
      properties:
        password:
          minLength: 1
          type: string
        username:
          minLength: 1
          type: string
      required:
      - password
      - username
      type: object
...
    RegisterRequest:
      properties:
        email:
          format: email
          minLength: 1
          type: string
        name:
          type: string
        password:
          maxLength: 2147483647
          minLength: 8
          type: string
        username:
          maxLength: 50
          minLength: 3
          type: string
      required:
      - email
      - password
      - username
      type: object
...
    TokenResponse:
      properties:
        token:
          type: string
      type: object
...
    UserResponse:
      properties:
        avatarInitials:
          type: string
        email:
          type: string
        id:
          type: string
        name:
          type: string
        username:
          type: string
      type: object
...
  securitySchemes:
    bearer-jwt:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  title: Personalised News Aggregator API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /api/content/articles/saved:
    post:
      description: Returns articles matching the provided list of IDs, useful for fetching a user's saved
        articles
      operationId: getSaved
      requestBody:
        content:
          application/json:
            schema:
              items:
                type: string
              type: array
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Article'
                type: array
          description: Matching articles
        '401':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Article'
                type: array
          description: Not authenticated
      security:
      - bearer-jwt: []
      summary: Batch-get articles by IDs
      tags:
      - Articles
  /api/content/articles/{id}:
    get:
      operationId: get_1
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
          description: Article found
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
          description: Article not found
      summary: Get full article
      tags:
      - Articles
...
  /api/users/auth/login:
    post:
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
          description: JWT token
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
          description: Invalid credentials
      summary: Authenticate and obtain a JWT
      tags:
      - Authentication
  /api/users/auth/register:
    post:
      description: Creates a new user account. The OAuth2 token endpoint at /oauth2/token handles login.
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
          description: User created
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
          description: Validation error or duplicate username/email
      summary: Register a new user
      tags:
      - Authentication
...
servers:
- url: /